move context extensions to extensions package.
