+++ app/src/main/java/wordtextcounter/details/main/App.kt
@@ -10,6 +10,7 @@ import wordtextcounter.details.main.store.ReportDatabase.Companion.DB_VERSION
 import wordtextcounter.details.main.store.migration.MigrationProcessor
 import wordtextcounter.details.main.util.Constants.PREF_DB_STORED
 import wordtextcounter.details.main.util.dbExists
+import wordtextcounter.details.main.util.extensions.getPreference
 
 class App : Application() {
